Add syntax highlighting to code blocks
This configures Redcarpet to use the Rouge gem for syntax highlighting.
It also moves the `ActionText::Markdown` class to lib, as it's
potentially part of the library, not the application domain.

You can still customize the markdown renderer by creating a custom renderer
class and setting it in the `ActionText::Markdown.renderer` attribute,
as illustrated in the `markdown.rb` initializer.

Fixes https://3.basecamp.com/2914079/buckets/36389984/card_tables/cards/7446319073